Reorganize helm code in helmutils package
